services/search/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ElasticSearch
elasticsearch==8.11.0

# HTTP Client (pour appels inter-services)
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0

